Restore shared journal ownership boundaries
Move content-addressed reserve and supply-attribution journal runtime schemas, hashing, constructors, and tests into shared/lib. Keep only the declarative reserve source-origin type in shared/types, and update Worker consumers so shared/types no longer depends on shared/lib.
