+++ worker/src/lib/__tests__/safety-score-v9-supply-attribution-journal-store.test.ts
@@ -5,7 +5,7 @@ import { describe, expect, it } from "vitest";
 import {
   createSupplyAttributionJournalV1,
   type SupplyAttributionJournalV1Payload,
-} from "@shared/types/safety-score-v9-supply-attribution-journal";
+} from "@shared/lib/safety-score-v9-supply-attribution-journal";
 import { createSqliteD1 } from "../../test-helpers/sqlite-d1";
 import {
   SUPPLY_ATTRIBUTION_JOURNAL_STORE_MAX_ROWS_PER_ASSET,
